Use `f32` for `scale_factor`

+++ devtools/src/lib.rs
@@ -98,7 +98,7 @@ where
         state.style(&self.program, theme)
     }
 
-    fn scale_factor(&self, state: &Self::State, window: window::Id) -> f64 {
+    fn scale_factor(&self, state: &Self::State, window: window::Id) -> f32 {
         state.scale_factor(&self.program, window)
     }
 }
@@ -397,7 +397,7 @@ where
         program.style(self.state(), theme)
     }
 
-    fn scale_factor(&self, program: &P, window: window::Id) -> f64 {
+    fn scale_factor(&self, program: &P, window: window::Id) -> f32 {
         program.scale_factor(self.state(), window)
     }
